Add images “checkIcon” and “deletedIcon” When productAmount is 0 it will be deleted

+++ FinalProject/FinalProject/FridgeCollectionViewController.h
@@ -41,7 +41,7 @@
 @property (strong, nonatomic) NSMutableArray *buttons;
 @property (nonatomic)int uiButtonX;
 @property (nonatomic)int countButton;
-@property (nonatomic)NSInteger clickedIndex;
+@property (nonatomic)int clickedIndex;
 @property (nonatomic)NSInteger amount;
 
 @property (nonatomic) int isEditProduct;
